+++ C++Projects/Bomb Dodger/Bomb Dodger/Tile.h
@@ -13,8 +13,8 @@ public:
 	// Constructor
 	Tile(size_t index, size_t columnNum, size_t rowNum, size_t colth, size_t rowth) :ObjBase()
 	{
-		x_size_ = GAME_WIDTH / (int)columnNum;
-		y_size_ = GAME_HEIGHT / (int)rowNum;
+		x_size_ = (float)GAME_WIDTH / (int)columnNum;
+		y_size_ = (float)GAME_HEIGHT / (int)rowNum;
 		x_ = float(colth * x_size_);
 		y_ = float(rowth * y_size_);
 	}
